refactor(gui): extract filter controls from menu bar
